WIP: Refactor authn/authz related code.
