+++ docs/authentication.rst
@@ -3,5 +3,8 @@ Authentication
 Authentication classes are used to associate a request with a user. Unless otherwise noted, all of the classes below
 adhere to the Django `REST Framework's API for authentication classes <http://www.django-rest-framework.org/api-guide/authentication/>`_.
 
-.. automodule:: edx_rest_framework_extensions.authentication
+.. automodule:: edx_rest_framework_extensions.auth.bearer.authentication
+    :members:
+
+.. automodule:: edx_rest_framework_extensions.auth.jwt.authentication
     :members:
