+++ edx_rest_framework_extensions/auth/README.rst
@@ -1,0 +1,5 @@
+This directory will contain all authentication and permission related extensions.
+
+It contains the following subdirectories:
+- bearer
+- jwt
