+++ edx_rest_framework_extensions/auth/jwt/constants.py
