+++ edx_rest_framework_extensions/auth/jwt/cookies.py
